+++ tour_server/src/main/java/com/dujiangyan/tour_server/entity/Detail.java
@@ -11,13 +11,13 @@ public class Detail {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private long id;
+    private int id;
 
     @Column(name = "orderId")
-    private long orderId;
+    private int orderId;
 
     @Column(name = "productId")
-    private long productId;
+    private int productId;
 
     @Column(name = "number")
     private int number;
